Update Web Platform Tests for storage APIs in 'data:' contexts
Per https://github.com/w3c/web-platform-tests/pull/4832 remove
WebSQL cases from WPT's workers/data-url*.html since it's not
otherwise tested in WPT, and the test "passes" even if the API is
not present. Also add an addition assertion that indexedDB is
present before blithely checking that using it throws.

WebSQL is not exposed to Workers in Chrome so we don't lose
coverage here - we only "passed" because the API was not present
in these contexts. The webexposed/ tests cover this.

Change-Id: I0aa05ff9e674351a2a49ba0bae8e5a9938b2cf34